AUXClient Target, TKSeparator Web Component
All static HTML sources have been moved to a separate AUXClient Target, preventing AUXLibrary resources from being leaked on the web and eliminating the need to run the TypeScript compiler when web-related sources have not been modified.
TKSeparator is the web component used in headers, lists and toolbars as a dividing line.