Fork of the WebAuthn-Swift library as the Validation class, HPBase64 class, fixed WebAuthn payloads being converted to Base64 instead of Base64URL, fixed TKClock marks not displaying correctly on browsers not supporting the CSS calc “pi” function
When using hardware tokens with Chromium based browsers, the extensionData flag of the attestationObject returned after invoking navigator.credentials.create is set to true even though no such data is provided, leading the library to reject such assertion as corrupted. My fork removes the raising of the exception since the extensionData is not used elsewhere in the library to validate credentials.